refactor(db/tests): extract shared apply_first_n migration helper
Both mod_tests_v18.rs and mod_tests_v19.rs shipped near-identical
apply_through_vN-1 functions whose only difference was the slice bound
(MIGRATIONS[..17] vs [..18]). Every future migration test would have
copy-pasted the same 30-line block.

Move the helper to a sibling mod_test_helpers.rs and parameterize on
the slice bound: apply_first_n(&pool, N-1). Add a column_names(pool,
table) helper to deduplicate the PRAGMA table_info fetch that
appeared 3× across V19 tests.

Drop the #![allow(unused_imports)] suppression — all imports in the
test files are now genuinely used.

Closes #103
