+++ crates/sp-server/src/db/mod.rs
@@ -329,6 +329,10 @@ pub async fn current_schema_version(pool: &SqlitePool) -> Result<i32, sqlx::Erro
     Ok(row.get("v"))
 }
 
+#[path = "mod_test_helpers.rs"]
+#[cfg(test)]
+mod test_helpers;
+
 #[path = "mod_tests.rs"]
 #[cfg(test)]
 mod tests;
